completed chapter 5

+++ shop/src/main/resources/db/migration/V1_2__create_product_table.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS product (
     id UUID NOT NULL PRIMARY KEY,
     name VARCHAR(20),
     description VARCHAR(128),
-    price double precision,
+    price numeric,
     weight double precision,
     category_id UUID REFERENCES productcategory(id),
     supplier varchar(20),
